artists splitted into base artist class and xyartist
